refactor(server): /api/jobs/:id/logs/stream + /api/events SSE 라우트 분할 (Plan C #C9 11단계)
- src/server/routes/sse.ts 신규 (142줄): 2개 SSE 라우트 캡슐화
  - GET /api/jobs/:id/logs/stream (단일 잡 로그 폴링)
  - GET /api/events (글로벌 잡/큐 상태)
- getInitialJobs 헬퍼 + SSE_INITIAL_JOB_LIMIT 상수를 라우트 모듈로 이전
- SSE 공용 헤더/타이밍 상수화 (SSE_HEADERS, SSE_MAX_DURATION_MS, *_POLL_MS)
- dashboard-api.ts에서 randomUUID import 제거
- dashboard-api.ts 521 → 402줄 (-119)

남은: auth (POST /api/auth + 미들웨어 — 다음 단계 또는 #C12와 함께)
